NCI-Agency/anet#2954: Put each aggregation widget in its own file
